requirements.txt
asgiref==3.5.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.0.14
django-cors-headers==3.11.0
finance-datareader==0.9.33
idna==3.3
lxml==4.9.0
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
